refactor(contentful): move code to createSchemaCustomization
