+++ packages/gatsby-source-contentful/src/gatsby-node.js
@@ -1,28 +1,20 @@
-const path = require(`path`)
-const isOnline = require(`is-online`)
+// @todo import syntax!
 const _ = require(`lodash`)
-const fs = require(`fs-extra`)
-const { createClient } = require(`contentful`)
-const v8 = require(`v8`)
 const fetch = require(`@vercel/fetch-retry`)(require(`node-fetch`))
-const { CODES } = require(`./report`)
 
-const normalize = require(`./normalize`)
-const fetchData = require(`./fetch`)
-const { createPluginConfig, maskText } = require(`./plugin-options`)
-const { downloadContentfulAssets } = require(`./download-contentful-assets`)
+const { createSchemaCustomization } = require(`./create-schema-customization`)
+const { sourceNodes } = require(`./source-nodes`)
+const { onPreBootstrap } = require(`./on-pre-bootstrap`)
+const { maskText } = require(`./plugin-options`)
 
-const conflictFieldPrefix = `contentful`
+exports.setFieldsOnGraphQLNodeType =
+  require(`./extend-node-type`).extendNodeType
 
-// restrictedNodeFields from here https://www.gatsbyjs.org/docs/node-interface/
-const restrictedNodeFields = [
-  `children`,
-  `contentful_id`,
-  `fields`,
-  `id`,
-  `internal`,
-  `parent`,
-]
+exports.createSchemaCustomization = createSchemaCustomization
+
+exports.sourceNodes = sourceNodes
+
+exports.onPreBootstrap = onPreBootstrap
 
 exports.setFieldsOnGraphQLNodeType =
   require(`./extend-node-type`).extendNodeType
@@ -149,597 +141,3 @@ List of locales and their codes can be found in Contentful app -> Settings -> Lo
     .external(validateContentfulAccess)
 
 exports.pluginOptionsSchema = pluginOptionsSchema
-
-/***
- * Localization algorithm
- *
- * 1. Make list of all resolvable IDs worrying just about the default ids not
- * localized ids
- * 2. Make mapping between ids, again not worrying about localization.
- * 3. When creating entries and assets, make the most localized version
- * possible for each localized node i.e. get the localized field if it exists
- * or the fallback field or the default field.
- */
-
-exports.sourceNodes = async (
-  {
-    actions,
-    getNode,
-    getNodes,
-    getNodesByType,
-    createNodeId,
-    store,
-    cache,
-    getCache,
-    reporter,
-    schema,
-    parentSpan,
-  },
-  pluginOptions
-) => {
-  const { createNode, deleteNode, touchNode, createTypes } = actions
-
-  let currentSyncData
-  let contentTypeItems
-  let tagItems
-  let defaultLocale
-  let locales
-  let space
-  if (process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE) {
-    reporter.info(
-      `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE: Storing/loading remote data through \`` +
-        process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE +
-        `\` so Remote changes CAN NOT be detected!`
-    )
-  }
-  const forceCache = await fs.exists(
-    process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE
-  )
-
-  const pluginConfig = createPluginConfig(pluginOptions)
-  const sourceId = `${pluginConfig.get(`spaceId`)}-${pluginConfig.get(
-    `environment`
-  )}`
-
-  const CACHE_SYNC_TOKEN = `contentful-sync-token-${sourceId}`
-  const CACHE_SYNC_DATA = `contentful-sync-data-${sourceId}`
-
-  /*
-   * Subsequent calls of Contentfuls sync API return only changed data.
-   *
-   * In some cases, especially when using rich-text fields, there can be data
-   * missing from referenced entries. This breaks the reference matching.
-   *
-   * To workround this, we cache the initial sync data and merge it
-   * with all data from subsequent syncs. Afterwards the references get
-   * resolved via the Contentful JS SDK.
-   */
-  const syncToken = await cache.get(CACHE_SYNC_TOKEN)
-  let previousSyncData = {
-    assets: [],
-    entries: [],
-  }
-  const cachedData = await cache.get(CACHE_SYNC_DATA)
-
-  if (cachedData) {
-    previousSyncData = cachedData
-  }
-
-  const fetchActivity = reporter.activityTimer(
-    `Contentful: Fetch data (${sourceId})`,
-    {
-      parentSpan,
-    }
-  )
-
-  if (forceCache) {
-    // If the cache has data, use it. Otherwise do a remote fetch anyways and prime the cache now.
-    // If present, do NOT contact contentful, skip the round trips entirely
-    reporter.info(
-      `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE was set. Skipping remote fetch, using data stored in \`${process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE}\``
-    )
-    ;({
-      currentSyncData,
-      contentTypeItems,
-      tagItems,
-      defaultLocale,
-      locales,
-      space,
-    } = v8.deserialize(
-      fs.readFileSync(process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE)
-    ))
-  } else {
-    const online = await isOnline()
-
-    // If the user knows they are offline, serve them cached result
-    // For prod builds though always fail if we can't get the latest data
-    if (
-      !online &&
-      process.env.GATSBY_CONTENTFUL_OFFLINE === `true` &&
-      process.env.NODE_ENV !== `production`
-    ) {
-      getNodes().forEach(node => {
-        if (node.internal.owner !== `gatsby-source-contentful`) {
-          return
-        }
-        touchNode(node)
-        if (node.localFile___NODE) {
-          // Prevent GraphQL type inference from crashing on this property
-          touchNode(getNode(node.localFile___NODE))
-        }
-      })
-
-      reporter.info(`Using Contentful Offline cache ⚠️`)
-      reporter.info(
-        `Cache may be invalidated if you edit package.json, gatsby-node.js or gatsby-config.js files`
-      )
-
-      return
-    }
-    if (process.env.GATSBY_CONTENTFUL_OFFLINE) {
-      reporter.info(
-        `Note: \`GATSBY_CONTENTFUL_OFFLINE\` was set but it either was not \`true\`, we _are_ online, or we are in production mode, so the flag is ignored.`
-      )
-    }
-
-    fetchActivity.start()
-    ;({
-      currentSyncData,
-      contentTypeItems,
-      tagItems,
-      defaultLocale,
-      locales,
-      space,
-    } = await fetchData({
-      syncToken,
-      reporter,
-      pluginConfig,
-      parentSpan,
-    }))
-
-    if (process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE) {
-      reporter.info(
-        `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE was set. Writing v8 serialized glob of remote data to: ` +
-          process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE
-      )
-      fs.writeFileSync(
-        process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE,
-        v8.serialize({
-          currentSyncData,
-          contentTypeItems,
-          defaultLocale,
-          locales,
-          space,
-        })
-      )
-    }
-  }
-
-  // Check for restricted content type names
-  const useNameForId = pluginConfig.get(`useNameForId`)
-  const restrictedContentTypes = [`entity`, `reference`, `asset`]
-
-  if (pluginConfig.get(`enableTags`)) {
-    restrictedContentTypes.push(`tags`)
-  }
-
-  contentTypeItems.forEach(contentTypeItem => {
-    // Establish identifier for content type
-    //  Use `name` if specified, otherwise, use internal id (usually a natural-language constant,
-    //  but sometimes a base62 uuid generated by Contentful, hence the option)
-    let contentTypeItemId
-    if (useNameForId) {
-      contentTypeItemId = contentTypeItem.name.toLowerCase()
-    } else {
-      contentTypeItemId = contentTypeItem.sys.id.toLowerCase()
-    }
-
-    if (restrictedContentTypes.includes(contentTypeItemId)) {
-      reporter.panic({
-        id: CODES.FetchContentTypes,
-        context: {
-          sourceMessage: `Restricted ContentType name found. The name "${contentTypeItemId}" is not allowed.`,
-        },
-      })
-    }
-  })
-
-  const allLocales = locales
-  locales = locales.filter(pluginConfig.get(`localeFilter`))
-  reporter.verbose(
-    `Default locale: ${defaultLocale}.   All locales: ${allLocales
-      .map(({ code }) => code)
-      .join(`, `)}`
-  )
-  if (allLocales.length !== locales.length) {
-    reporter.verbose(
-      `After plugin.options.localeFilter: ${locales
-        .map(({ code }) => code)
-        .join(`, `)}`
-    )
-  }
-  if (locales.length === 0) {
-    reporter.panic({
-      id: CODES.LocalesMissing,
-      context: {
-        sourceMessage: `Please check if your localeFilter is configured properly. Locales '${allLocales
-          .map(item => item.code)
-          .join(`,`)}' were found but were filtered down to none.`,
-      },
-    })
-  }
-
-  if (pluginConfig.get(`enableTags`)) {
-    createTypes(
-      schema.buildObjectType({
-        name: `ContentfulTag`,
-        fields: {
-          name: { type: `String!` },
-          contentful_id: { type: `String!` },
-          id: { type: `ID!` },
-        },
-        interfaces: [`Node`],
-        extensions: { dontInfer: {} },
-      })
-    )
-  }
-
-  createTypes(`
-  interface ContentfulEntry implements Node {
-    contentful_id: String!
-    id: ID!
-    node_locale: String!
-  }
-`)
-
-  createTypes(`
-  interface ContentfulReference {
-    contentful_id: String!
-    id: ID!
-  }
-`)
-
-  createTypes(
-    schema.buildObjectType({
-      name: `ContentfulAsset`,
-      fields: {
-        contentful_id: { type: `String!` },
-        id: { type: `ID!` },
-      },
-      interfaces: [`ContentfulReference`, `Node`],
-    })
-  )
-
-  const gqlTypes = contentTypeItems.map(contentTypeItem =>
-    schema.buildObjectType({
-      name: _.upperFirst(
-        _.camelCase(
-          `Contentful ${
-            pluginConfig.get(`useNameForId`)
-              ? contentTypeItem.name
-              : contentTypeItem.sys.id
-          }`
-        )
-      ),
-      fields: {
-        contentful_id: { type: `String!` },
-        id: { type: `ID!` },
-        node_locale: { type: `String!` },
-      },
-      interfaces: [`ContentfulReference`, `ContentfulEntry`, `Node`],
-    })
-  )
-
-  createTypes(gqlTypes)
-
-  fetchActivity.end()
-
-  const processingActivity = reporter.activityTimer(
-    `Contentful: Process data (${sourceId})`,
-    {
-      parentSpan,
-    }
-  )
-  processingActivity.start()
-
-  // Create a map of up to date entries and assets
-  function mergeSyncData(previous, current, deleted) {
-    const entryMap = new Map()
-    previous.forEach(e => !deleted.has(e.sys.id) && entryMap.set(e.sys.id, e))
-    current.forEach(e => !deleted.has(e.sys.id) && entryMap.set(e.sys.id, e))
-    return [...entryMap.values()]
-  }
-
-  const deletedSet = new Set(currentSyncData.deletedEntries.map(e => e.sys.id))
-
-  const mergedSyncData = {
-    entries: mergeSyncData(
-      previousSyncData.entries,
-      currentSyncData.entries,
-      deletedSet
-    ),
-    assets: mergeSyncData(
-      previousSyncData.assets,
-      currentSyncData.assets,
-      deletedSet
-    ),
-  }
-
-  // Store a raw and unresolved copy of the data for caching
-  const mergedSyncDataRaw = _.cloneDeep(mergedSyncData)
-
-  // Use the JS-SDK to resolve the entries and assets
-  const res = createClient({
-    space: `none`,
-    accessToken: `fake-access-token`,
-  }).parseEntries({
-    items: mergedSyncData.entries,
-    includes: {
-      assets: mergedSyncData.assets,
-      entries: mergedSyncData.entries,
-    },
-  })
-
-  mergedSyncData.entries = res.items
-
-  // Inject raw API output to rich text fields
-  const richTextFieldMap = new Map()
-  contentTypeItems.forEach(contentType => {
-    richTextFieldMap.set(
-      contentType.sys.id,
-      contentType.fields
-        .filter(field => field.type === `RichText`)
-        .map(field => field.id)
-    )
-  })
-
-  const rawEntries = new Map()
-  mergedSyncDataRaw.entries.forEach(rawEntry =>
-    rawEntries.set(rawEntry.sys.id, rawEntry)
-  )
-
-  mergedSyncData.entries.forEach(entry => {
-    const contentTypeId = entry.sys.contentType.sys.id
-    const richTextFieldIds = richTextFieldMap.get(contentTypeId)
-    if (richTextFieldIds) {
-      richTextFieldIds.forEach(richTextFieldId => {
-        if (!entry.fields[richTextFieldId]) {
-          return
-        }
-        entry.fields[richTextFieldId] = rawEntries.get(entry.sys.id).fields[
-          richTextFieldId
-        ]
-      })
-    }
-  })
-
-  const entryList = normalize.buildEntryList({
-    mergedSyncData,
-    contentTypeItems,
-  })
-
-  // Remove deleted entries & assets.
-  // TODO figure out if entries referencing now deleted entries/assets
-  // are "updated" so will get the now deleted reference removed.
-
-  function deleteContentfulNode(node) {
-    const normalizedType = node.sys.type.startsWith(`Deleted`)
-      ? node.sys.type.substring(`Deleted`.length)
-      : node.sys.type
-
-    const localizedNodes = locales
-      .map(locale => {
-        const nodeId = createNodeId(
-          normalize.makeId({
-            spaceId: space.sys.id,
-            id: node.sys.id,
-            type: normalizedType,
-            currentLocale: locale.code,
-            defaultLocale,
-          })
-        )
-        return getNode(nodeId)
-      })
-      .filter(node => node)
-
-    localizedNodes.forEach(node => {
-      // touchNode first, to populate typeOwners & avoid erroring
-      touchNode(node)
-      deleteNode(node)
-    })
-  }
-
-  currentSyncData.deletedEntries.forEach(deleteContentfulNode)
-  currentSyncData.deletedAssets.forEach(deleteContentfulNode)
-
-  const existingNodes = getNodes().filter(
-    n =>
-      n.internal.owner === `gatsby-source-contentful` &&
-      (n?.sys?.type === `Asset` || n?.sys?.type === `Entry`)
-  )
-  existingNodes.forEach(n => touchNode(n))
-
-  const assets = mergedSyncData.assets
-
-  reporter.info(`Updated entries ${currentSyncData.entries.length}`)
-  reporter.info(`Deleted entries ${currentSyncData.deletedEntries.length}`)
-  reporter.info(`Updated assets ${currentSyncData.assets.length}`)
-  reporter.info(`Deleted assets ${currentSyncData.deletedAssets.length}`)
-
-  // Update syncToken
-  const nextSyncToken = currentSyncData.nextSyncToken
-
-  await Promise.all([
-    cache.set(CACHE_SYNC_DATA, mergedSyncDataRaw),
-    cache.set(CACHE_SYNC_TOKEN, nextSyncToken),
-  ])
-
-  reporter.verbose(`Building Contentful reference map`)
-
-  // Create map of resolvable ids so we can check links against them while creating
-  // links.
-  const resolvable = normalize.buildResolvableSet({
-    existingNodes,
-    entryList,
-    assets,
-    defaultLocale,
-    locales,
-    space,
-  })
-
-  // Build foreign reference map before starting to insert any nodes
-  const foreignReferenceMap = normalize.buildForeignReferenceMap({
-    contentTypeItems,
-    entryList,
-    resolvable,
-    defaultLocale,
-    locales,
-    space,
-    useNameForId: pluginConfig.get(`useNameForId`),
-  })
-
-  reporter.verbose(`Resolving Contentful references`)
-
-  const newOrUpdatedEntries = new Set()
-  entryList.forEach(entries => {
-    entries.forEach(entry => {
-      newOrUpdatedEntries.add(`${entry.sys.id}___${entry.sys.type}`)
-    })
-  })
-
-  // Update existing entry nodes that weren't updated but that need reverse
-  // links added.
-  existingNodes
-    .filter(n => newOrUpdatedEntries.has(`${n.id}___${n.sys.type}`))
-    .forEach(n => {
-      if (foreignReferenceMap[`${n.id}___${n.sys.type}`]) {
-        foreignReferenceMap[`${n.id}___${n.sys.type}`].forEach(
-          foreignReference => {
-            // Add reverse links
-            if (n[foreignReference.name]) {
-              n[foreignReference.name].push(foreignReference.id)
-              // It might already be there so we'll uniquify after pushing.
-              n[foreignReference.name] = _.uniq(n[foreignReference.name])
-            } else {
-              // If is one foreign reference, there can always be many.
-              // Best to be safe and put it in an array to start with.
-              n[foreignReference.name] = [foreignReference.id]
-            }
-          }
-        )
-      }
-    })
-
-  processingActivity.end()
-
-  const creationActivity = reporter.activityTimer(
-    `Contentful: Create nodes (${sourceId})`,
-    {
-      parentSpan,
-    }
-  )
-  creationActivity.start()
-
-  for (let i = 0; i < contentTypeItems.length; i++) {
-    const contentTypeItem = contentTypeItems[i]
-
-    if (entryList[i].length) {
-      reporter.info(
-        `Creating ${entryList[i].length} Contentful ${
-          pluginConfig.get(`useNameForId`)
-            ? contentTypeItem.name
-            : contentTypeItem.sys.id
-        } nodes`
-      )
-    }
-
-    // A contentType can hold lots of entries which create nodes
-    // We wait until all nodes are created and processed until we handle the next one
-    // TODO add batching in gatsby-core
-    await Promise.all(
-      normalize.createNodesForContentType({
-        contentTypeItem,
-        restrictedNodeFields,
-        conflictFieldPrefix,
-        entries: entryList[i],
-        createNode,
-        createNodeId,
-        getNode,
-        resolvable,
-        foreignReferenceMap,
-        defaultLocale,
-        locales,
-        space,
-        useNameForId: pluginConfig.get(`useNameForId`),
-        pluginConfig,
-      })
-    )
-  }
-
-  if (assets.length) {
-    reporter.info(`Creating ${assets.length} Contentful Asset nodes`)
-  }
-
-  for (let i = 0; i < assets.length; i++) {
-    // We wait for each asset to be process until handling the next one.
-    await Promise.all(
-      normalize.createAssetNodes({
-        assetItem: assets[i],
-        createNode,
-        createNodeId,
-        defaultLocale,
-        locales,
-        space,
-      })
-    )
-  }
-
-  // Create tags entities
-  if (tagItems.length) {
-    reporter.info(`Creating ${tagItems.length} Contentful Tag nodes`)
-
-    for (const tag of tagItems) {
-      await createNode({
-        id: createNodeId(`ContentfulTag__${space.sys.id}__${tag.sys.id}`),
-        name: tag.name,
-        contentful_id: tag.sys.id,
-        internal: {
-          type: `ContentfulTag`,
-          contentDigest: tag.sys.updatedAt,
-        },
-      })
-    }
-  }
-
-  creationActivity.end()
-
-  if (pluginConfig.get(`downloadLocal`)) {
-    reporter.info(`Download Contentful asset files`)
-
-    await downloadContentfulAssets({
-      actions,
-      createNodeId,
-      store,
-      cache,
-      getCache,
-      getNode,
-      getNodesByType,
-      reporter,
-      assetDownloadWorkers: pluginConfig.get(`assetDownloadWorkers`),
-    })
-  }
-
-  return
-}
-
-// Check if there are any ContentfulAsset nodes and if gatsby-image is installed. If so,
-// add fragments for ContentfulAsset and gatsby-image. The fragment will cause an error
-// if there's not ContentfulAsset nodes and without gatsby-image, the fragment is useless.
-exports.onPreExtractQueries = async ({ store }) => {
-  const program = store.getState().program
-
-  const CACHE_DIR = path.resolve(
-    `${program.directory}/.cache/contentful/assets/`
-  )
-  await fs.ensureDir(CACHE_DIR)
-}
